requirements.txt
streamlit>=1.33.0
pymupdf>=1.23.8
python-docx>=1.1.0
pandas>=2.2.0
rapidfuzz>=3.6.1
openpyxl>=3.1.2
XlsxWriter>=3.1.2
